revert warning move
